+++ Source/Units/IronFrame.cpp
@@ -832,9 +832,9 @@ bool terFrame::canTeleportate() const
 
 bool terFrame::analyzeTerrain()
 {
-	int D = clamp(round(Vect2f(BodyPoint->boxMax()).norm()) >> kmGrid, 1, 1000);
-	int x0 = round(position().x) >> kmGrid;
-	int y0 = round(position().y) >> kmGrid;
+	int D = clamp((int)round(Vect2f(BodyPoint->boxMax()).norm()) >> kmGrid, 1, 1000);
+	int x0 = (int)round(position().x) >> kmGrid;
+	int y0 = (int)round(position().y) >> kmGrid;
 	int z0 = vMap.hZeroPlast;
 	int dzMax = attr().heightMax;
 	for(int y = -D; y <= D; y++)
